refactor options and require into the runtime library
